Basic first version.
If a commit message includes the hash #reviewthis, any github user
(signified by @username) will get an email (if they have a public
email address). Also, any email address includes gets an email as
well.

Currently, sendmail is used in dev, and since we're going to deploy
to heroku, the sendgrid addon is used in production.